requirements.txt
asgiref==3.10.0
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
django-jazzmin==3.0.1
djangorestframework==3.16.1
pillow==12.1.0
psycopg2-binary==2.9.11
python-dotenv==1.1.1
sqlparse==0.5.3
whitenoise==6.8.2
gunicorn==23.0.0
